Started work on Arena class that spawns enemies after enemies have been killed

+++ core/src/com/byrjamin/wickedwizard/MainGame.java
@@ -24,8 +24,8 @@ import com.byrjamin.wickedwizard.screens.PlayScreen;
 
 public class MainGame extends Game {
 
-	public static final int GAME_HEIGHT = 1200;
-	public static final int GAME_WIDTH = 1920;
+	public static final float GAME_HEIGHT = 1200;
+	public static final float GAME_WIDTH = 1920;
 
 
 	public static final int GAME_UNITS = 20;
